rebase best (#27)
* Add .circleci/config.yml (#9)

* feat: upgrade @types/node from 20.17.23 to 22.13.4 (#13)

Snyk has created this PR to upgrade @types/node from 20.17.23 to 22.13.4.

See this package in npm:
@types/node

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/467ac4e6-c558-42b4-8665-6455b882a79e?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* feat: upgrade framer-motion from 9.1.7 to 12.4.2 (#10)

Snyk has created this PR to upgrade framer-motion from 9.1.7 to 12.4.2.

See this package in npm:
framer-motion

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/467ac4e6-c558-42b4-8665-6455b882a79e?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* feat: upgrade react-dom from 18.3.1 to 19.0.0 (#12)

Snyk has created this PR to upgrade react-dom from 18.3.1 to 19.0.0.

See this package in npm:
react-dom

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/467ac4e6-c558-42b4-8665-6455b882a79e?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade 0xsequence from 2.2.3 to 2.2.13 (#14)

Snyk has created this PR to upgrade 0xsequence from 2.2.3 to 2.2.13.

See this package in npm:
0xsequence

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/467ac4e6-c558-42b4-8665-6455b882a79e?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* Create SECURITY.md (#15)

Signed-off-by: AU_gdev_19 <64915515+Dargon789@users.noreply.github.com>

* feat: upgrade @0xsequence/design-system from 1.9.0 to 2.0.2 (#19)

Snyk has created this PR to upgrade @0xsequence/design-system from 1.9.0 to 2.0.2.

See this package in npm:
@0xsequence/design-system

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/467ac4e6-c558-42b4-8665-6455b882a79e?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* feat: upgrade @types/react from 18.3.18 to 19.0.10 (#17)

Snyk has created this PR to upgrade @types/react from 18.3.18 to 19.0.10.

See this package in npm:
@types/react

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/467ac4e6-c558-42b4-8665-6455b882a79e?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* feat: upgrade react from 18.3.1 to 19.0.0 (#16)

Snyk has created this PR to upgrade react from 18.3.1 to 19.0.0.

See this package in npm:
react

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/467ac4e6-c558-42b4-8665-6455b882a79e?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade @0xsequence/abi from 2.2.13 to 2.2.14 (#21)

Snyk has created this PR to upgrade @0xsequence/abi from 2.2.13 to 2.2.14.

See this package in npm:
@0xsequence/abi

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/467ac4e6-c558-42b4-8665-6455b882a79e?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade @0xsequence/utils from 2.2.13 to 2.2.14 (#22)

Snyk has created this PR to upgrade @0xsequence/utils from 2.2.13 to 2.2.14.

See this package in npm:
@0xsequence/utils

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/467ac4e6-c558-42b4-8665-6455b882a79e?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* fix: upgrade @0xsequence/network from 2.2.13 to 2.2.14 (#23)

Snyk has created this PR to upgrade @0xsequence/network from 2.2.13 to 2.2.14.

See this package in npm:
@0xsequence/network

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/467ac4e6-c558-42b4-8665-6455b882a79e?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

* feat: upgrade typescript from 4.9.5 to 5.8.2 (#26)

Snyk has created this PR to upgrade typescript from 4.9.5 to 5.8.2.

See this package in npm:
typescript

See this project in Snyk:
https://app.snyk.io/org/dargon789/project/467ac4e6-c558-42b4-8665-6455b882a79e?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

---------

Signed-off-by: AU_gdev_19 <64915515+Dargon789@users.noreply.github.com>
Co-authored-by: snyk-bot <snyk-bot@snyk.io>